backend/app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile backend/app/requirements.in
#
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via starlette
bcrypt==4.0.1
    # via passlib
black==23.7.0
    # via -r backend/app/requirements.in
build==0.10.0
    # via pip-tools
cffi==1.15.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
click==8.1.6
    # via
    #   black
    #   pip-tools
    #   uvicorn
couchbase==4.1.8
    # via -r backend/app/requirements.in
cryptography==41.0.3
    # via python-jose
distlib==0.3.7
    # via virtualenv
ecdsa==0.18.0
    # via python-jose
fastapi==0.101.0
    # via -r backend/app/requirements.in
filelock==3.12.2
    # via virtualenv
h11==0.14.0
    # via uvicorn
identify==2.5.26
    # via pre-commit
idna==3.4
    # via anyio
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
packaging==23.1
    # via
    #   black
    #   build
passlib[bcrypt]==1.7.4
    # via -r backend/app/requirements.in
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via -r backend/app/requirements.in
platformdirs==3.10.0
    # via
    #   black
    #   virtualenv
pre-commit==3.3.3
    # via -r backend/app/requirements.in
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==2.1.1
    # via
    #   -r backend/app/requirements.in
    #   fastapi
pydantic-core==2.4.0
    # via pydantic
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r backend/app/requirements.in
python-jose[cryptography]==3.3.0
    # via -r backend/app/requirements.in
pyyaml==6.0.1
    # via pre-commit
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
typing-extensions==4.7.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.23.2
    # via -r backend/app/requirements.in
virtualenv==20.24.3
    # via pre-commit
wheel==0.41.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
